# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Grooming CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all grooming businesses
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a grooming business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Clients
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Service Packages
...
  /reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [reports]
      summary: Revenue report by period
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
  /reports/appointments:
    get:
      operationId: getAppointmentsReport
      tags: [reports]
      summary: Appointments report
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Appointments report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentsReport"
  /reports/top-clients:
    get:
      operationId: getTopClientsReport
      tags: [reports]
      summary: Top clients by revenue and frequency
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopClient"

  # Dashboard
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary stats
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
components:
...
  schemas:
...
    Tenant:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TenantInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string

    # Client
    Client:
      type: object
      required: [id, tenantId, name, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Pet:
      type: object
      required: [id, clientId, name, size, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        name:
          type: string
        breed:
          type: ["string", "null"]
        size:
          type: string
          enum:
            - mini_curto
            - mini_longo
            - pequeno_curto
            - pequeno_longo
            - medio_curto
            - medio_longo
            - grande_curto
            - grande_longo
            - gigante
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Service:
      type: object
      required: [id, tenantId, name, size, price, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        size:
          type: string
          enum:
            - mini_curto
            - mini_longo
            - pequeno_curto
            - pequeno_longo
            - medio_curto
            - medio_longo
            - grande_curto
            - grande_longo
            - gigante
        price:
          type: number
        durationMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ServiceInput:
      type: object
      required: [tenantId, name, size, price]
      properties:
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: string
        size:
          type: string
          enum:
            - mini_curto
            - mini_longo
            - pequeno_curto
            - pequeno_longo
            - medio_curto
            - medio_longo
            - grande_curto
            - grande_longo
            - gigante
        price:
          type: number
        durationMinutes:
          type: integer
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        size:
          type: string
          enum:
            - mini_curto
            - mini_longo
            - pequeno_curto
            - pequeno_longo
            - medio_curto
            - medio_longo
            - grande_curto
            - grande_longo
            - gigante
        price:
          type: number
        durationMinutes:
          type: integer

    # Package
...
    AppointmentFull:
      type: object
      required: [id, tenantId, petId, clientId, scheduledDate, status, totalPrice, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        petId:
          type: integer
        clientId:
          type: integer
        serviceId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        scheduledDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [aguardando, em_atendimento, concluido, cancelado]
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        recurringWeeks:
          type: ["integer", "null"]
        recurringGroupId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        pet:
          $ref: "#/components/schemas/Pet"
        client:
          $ref: "#/components/schemas/Client"
        service:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
            price:
              type: number
        package:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
            price:
              type: number
...
    RevenueReport:
      type: object
      required: [totalRevenue, totalAppointments, averageTicket, data]
      properties:
        totalRevenue:
          type: number
        totalAppointments:
          type: integer
        averageTicket:
          type: number
        data:
          type: array
          items:
            type: object
            required: [period, revenue, count]
            properties:
              period:
                type: string
              revenue:
                type: number
              count:
                type: integer
    AppointmentsReport:
      type: object
      required: [total, byStatus, bySize, data]
      properties:
        total:
          type: integer
        byStatus:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status:
                type: string
              count:
                type: integer
        bySize:
          type: array
          items:
            type: object
            required: [size, count]
            properties:
              size:
                type: string
              count:
                type: integer
        data:
          type: array
          items:
            type: object
            required: [date, count]
            properties:
              date:
                type: string
              count:
                type: integer
    TopClient:
      type: object
      required: [clientId, clientName, totalRevenue, totalAppointments]
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        totalRevenue:
          type: number
        totalAppointments:
          type: integer
        lastAppointmentDate:
          type: ["string", "null"]

    # Auth
...
    DashboardSummary:
      type: object
      required: [todayAppointments, weekAppointments, monthRevenue, activeClients, pendingAppointments]
      properties:
        todayAppointments:
          type: integer
        weekAppointments:
          type: integer
        monthRevenue:
          type: number
        activeClients:
          type: integer
        pendingAppointments:
          type: integer
        recentAppointments:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentFull"